requirements.txt
obs-websocket-py
pyserial
easygui
win11toast
